SkinGuardian/requirements.txt
flask==2.2.3
onnxruntime==1.14.1
numpy==1.23.5
pillow==9.4.0
